refactor(daemon): merge the split launch-planning layer into src/launch/
The launch planning was split across two directories for no reason other
than history: acp/runtime-launch.ts prepared the launch and
runtimes/launch-policy.ts composed on top of it, importing back across
the seam. Reunite them as src/launch/prepare.ts and src/launch/compose.ts
so the cross-directory edge becomes an intra-directory one.

Pure moves — no symbol renames, no behavior change. acp/spawn-driver.ts
and acp/sandbox.ts stay in acp/.
