+++ packages/daemon/src/cli/chat.ts
@@ -10,14 +10,14 @@ import { memoryKindOf, memoryProviderFor } from '../memory/provider.js'
 import { WorkspaceManager } from '../workspace/workspace-manager.js'
 import { configureWorkspaceGitOrigins } from '../workspace/git-origin-policy.js'
 import { AcpHost } from '../acp/acp-host.js'
-import { effectiveRunInSandbox } from '../acp/runtime-launch.js'
+import { effectiveRunInSandbox } from '../launch/prepare.js'
 import { detectSandbox } from '../acp/sandbox.js'
 import { resolveRoot } from '../paths.js'
 import { runtimeHomePath } from '../runtimes/runtime-home.js'
 import { installedRuntimeCatalog } from '../runtimes/probe.js'
 import { probeAllRuntimes, type ProbeOptions, type RuntimeProbeResult } from '../runtimes/runtime-prober.js'
 import { CuratedRuntimeAdmission } from '../runtimes/curated-admission.js'
-import { composeRuntimeLaunch } from '../runtimes/launch-policy.js'
+import { composeRuntimeLaunch } from '../launch/compose.js'
 import type { RuntimeDef } from '../config/config-schema.js'
 import { persistSkillSandboxRequirement } from '../skills/skill-sandbox-policy.js'
