+++ packages/daemon/src/evaluation/runner.ts
@@ -15,13 +15,13 @@ import { basename, join, resolve } from 'node:path'
 import { z } from 'zod'
 import { AcpHost, turnFailureCode } from '../acp/acp-host.js'
 import { detectSandbox } from '../acp/sandbox.js'
-import { effectiveRunInSandbox } from '../acp/runtime-launch.js'
+import { effectiveRunInSandbox } from '../launch/prepare.js'
 import { agentChildEnv } from '../agents/agent-env.js'
 import { selectAgent } from '../agents/load-agents.js'
 import { memoryProviderFor } from '../memory/provider.js'
 import { loadConfig } from '../config/load-config.js'
 import { Daemon, type DaemonEvaluationOptions, type DaemonEvaluationTurnResult } from '../daemon.js'
-import { composeRuntimeLaunch } from '../runtimes/launch-policy.js'
+import { composeRuntimeLaunch } from '../launch/compose.js'
 import { persistSkillSandboxRequirement } from '../skills/skill-sandbox-policy.js'
 import { DAEMON_VERSION } from '../version.js'
 import {
