+++ packages/daemon/test/acp-matrix/acp-matrix.test.ts
@@ -22,7 +22,7 @@ import { loadConfig } from '../../src/config/load-config.js'
 import type { RuntimeDef } from '../../src/config/config-schema.js'
 import { resolveRoot } from '../../src/paths.js'
 import { installSkills } from '../../src/skills/install-skills.js'
-import { composeRuntimeLaunch } from '../../src/runtimes/launch-policy.js'
+import { composeRuntimeLaunch } from '../../src/launch/compose.js'
 import { installedRuntimeCatalog } from '../../src/runtimes/probe.js'
 import { resolveRuntimeCatalog, type ResolvedRuntimeEntry } from '../../src/runtimes/registry.js'
 import { FEATURES, type FeatureId } from './support-matrix.js'
