+++ packages/daemon/test/runtime-launch.test.ts
@@ -15,8 +15,8 @@ import {
 import { tmpdir } from 'node:os'
 import { delimiter, dirname, isAbsolute, join, relative, sep } from 'node:path'
 import { parse as parseYaml } from 'yaml'
-import { prepareRuntimeLaunch } from '../src/acp/runtime-launch.js'
-import { composeRuntimeLaunch, runtimeSandboxReadRoots } from '../src/runtimes/launch-policy.js'
+import { prepareRuntimeLaunch } from '../src/launch/prepare.js'
+import { composeRuntimeLaunch, runtimeSandboxReadRoots } from '../src/launch/compose.js'
 import { runtimeMemoryCapabilities } from '../src/memory/runtime/capabilities.js'
 import { MemoryProviderUnavailableError } from '../src/memory/provider.js'
 import type { RuntimeDef } from '../src/config/config-schema.js'
